Make Java plugin an optional dependency
