+++ plugin/src/main/resources/META-INF/com.github.tarcv.testingteam.surveyoridea-withJava.xml
@@ -1,0 +1,3 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.intellij"/>
+</idea-plugin>
