+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -11,7 +11,7 @@
     <depends>com.intellij.modules.xml</depends>
     <!-- Modules/plugins that are optional or not present in all IDE: -->
     <!-- (This list should match platformPlugins in gradle.properties) -->
-    <depends>com.intellij.java</depends>
+    <depends optional="true" config-file="com.github.tarcv.testingteam.surveyoridea-withJava.xml">com.intellij.java</depends>
 
     <extensions defaultExtensionNs="com.intellij">
         <toolWindow id="Locate Element" anchor="bottom" secondary="false"
